// generated/src/app/api-models/price-rule-master.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class PriceRuleMaster {
  priceRule: string;
  description: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  drgGrouperType: string;
  drgPricerType: string;
  customTableLogic: string;
  currAnesMinUnit: number;
  currConvEffDate: Date;
  priorAnesMinUnit: number;
  assignDrg: string;
  useExtDrgPricer: string;
  fromMin1: number;
  thruMin1: number;
  currAnesMinUnit2: number;
  priorAnesMinUnit2: number;
  fromMin2: number;
  thruMin2: number;
  currAnesMinUnit3: number;
  priorAnesMinUnit3: number;
  fromMin3: number;
  thruMin3: number;
  currAnesMinUnit4: number;
  priorAnesMinUnit4: number;
  fromMin4: number;
  thruMin4: number;
  minimumMinForUnit: number;
  procedurePriceOption: string;
  modifierCode: string;
}